abstracted frame processing outside of the camera implementation
