+++ camerakit/src/main/api16/com/wonderkiln/camerakit/Camera1.java
@@ -1,6 +1,5 @@
 package com.wonderkiln.camerakit;
 
-import android.graphics.ImageFormat;
 import android.graphics.Rect;
 import android.graphics.YuvImage;
 import android.hardware.Camera;
@@ -16,18 +15,16 @@ import android.view.MotionEvent;
 import android.view.View;
 
 import com.google.android.gms.vision.Detector;
-import com.google.android.gms.vision.Frame;
+import com.google.android.gms.vision.text.TextBlock;
+import com.wonderkiln.camerakit.textrecognition.FrameProcessingRunnable;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -82,20 +79,14 @@ public class Camera1 extends CameraImpl {
     @VideoQuality
     private int mVideoQuality;
 
+    private Detector<TextBlock> mTextDetector;
+
     private int mVideoBitRate;
 
     private Handler mainHandler = new Handler(Looper.getMainLooper());
     private Handler mHandler = new Handler();
 
-    private Thread mProcessingThread;
     private FrameProcessingRunnable mFrameProcessor;
-
-    /**
-     * Map to convert between a byte array, received from the camera, and its associated byte
-     * buffer.  We use byte buffers internally because this is a more efficient way to call into
-     * native code later (avoids a potential copy).
-     */
-    private Map<byte[], ByteBuffer> mBytesToByteBuffer = new HashMap<>();
 
     private VideoCapturedCallback mVideoCallback;
 
@@ -127,15 +118,6 @@ public class Camera1 extends CameraImpl {
         mCameraInfo = new Camera.CameraInfo();
     }
 
-    Camera1(EventDispatcher eventDispatcher, PreviewImpl preview, Detector<?> detector){
-        this(eventDispatcher, preview);
-        mFrameProcessor = new com.wonderkiln.camerakit.Camera1.FrameProcessingRunnable(detector);
-        // start text detection thread
-        mProcessingThread = new Thread(mFrameProcessor);
-        mFrameProcessor.setActive(true);
-        mProcessingThread.start();
-    }
-
     // CameraImpl:
 
     @Override
@@ -164,22 +146,9 @@ public class Camera1 extends CameraImpl {
 
         releaseMediaRecorder();
         releaseCamera();
-
-        // stop text dectection thread
-        if (mProcessingThread != null) {
-            try {
-                // Wait for the thread to complete to ensure that we can't have multiple threads
-                // executing at the same time (i.e., which would happen if we called start too
-                // quickly after stop).
-                mProcessingThread.join();
-            } catch (InterruptedException e) {
-                Log.d(TAG, "Frame processing thread interrupted on release.");
-            }
-            mProcessingThread = null;
+        if (mFrameProcessor != null) {
+            mFrameProcessor.cleanup();
         }
-
-        mFrameProcessor.setActive(false);
-        mBytesToByteBuffer.clear();
     }
 
     void setDisplayAndDeviceOrientation() {
@@ -296,6 +265,11 @@ public class Camera1 extends CameraImpl {
     @Override
     void setZoom(@Zoom int zoom) {
         this.mZoom = zoom;
+    }
+
+    @Override
+    void setTextDetector(Detector<TextBlock> detector) {
+        this.mTextDetector = detector;
     }
 
     @Override
@@ -555,16 +529,12 @@ public class Camera1 extends CameraImpl {
             adjustCameraParameters();
 
             mEventDispatcher.dispatch(new CameraKitEvent(CameraKitEvent.TYPE_CAMERA_OPEN));
-            mCamera.setPreviewCallbackWithBuffer(new Camera.PreviewCallback() {
-                @Override
-                public void onPreviewFrame(byte[] bytes, Camera camera) {
-                    mFrameProcessor.setNextFrame(bytes, camera);
-                }
-            });
-            mCamera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
-            mCamera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
-            mCamera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
-            mCamera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
+
+            if (mTextDetector != null) {
+                // start text detection thread
+                mFrameProcessor = new FrameProcessingRunnable(mTextDetector, mPreviewSize, mCamera);
+                mFrameProcessor.start();
+            }
         }
     }
 
@@ -591,38 +561,11 @@ public class Camera1 extends CameraImpl {
                 mVideoSize = null;
 
                 mEventDispatcher.dispatch(new CameraKitEvent(CameraKitEvent.TYPE_CAMERA_CLOSE));
-                mFrameProcessor.release();
+                if (mFrameProcessor != null) {
+                    mFrameProcessor.release();
+                }
             }
         }
-    }
-
-    /**
-     * Creates one buffer for the camera preview callback.  The size of the buffer is based off of
-     * the camera preview size and the format of the camera image.
-     *
-     * @return a new preview buffer of the appropriate size for the current camera settings
-     */
-    private byte[] createPreviewBuffer(Size previewSize) {
-        int bitsPerPixel = ImageFormat.getBitsPerPixel(ImageFormat.NV21);
-        long sizeInBits = previewSize.getHeight() * previewSize.getWidth() * bitsPerPixel;
-        int bufferSize = (int) Math.ceil(sizeInBits / 8.0d) + 1;
-
-        //
-        // NOTICE: This code only works when using play services v. 8.1 or higher.
-        //
-
-        // Creating the byte array this way and wrapping it, as opposed to using .allocate(),
-        // should guarantee that there will be an array to work with.
-        byte[] byteArray = new byte[bufferSize];
-        ByteBuffer buffer = ByteBuffer.wrap(byteArray);
-        if (!buffer.hasArray() || (buffer.array() != byteArray)) {
-            // I don't think that this will ever happen.  But if it does, then we wouldn't be
-            // passing the preview content to the underlying detector later.
-            throw new IllegalStateException("Failed to create valid buffer for camera source.");
-        }
-
-        mBytesToByteBuffer.put(byteArray, buffer);
-        return byteArray;
     }
 
     private int calculatePreviewRotation() {
@@ -755,18 +698,6 @@ public class Camera1 extends CameraImpl {
         } catch (Exception e) {
             notifyErrorListener(e);
         }
-//
-//        SizePair sizePair = selectSizePair(mCamera, 1024, 768);
-//        if (sizePair == null) {
-//            throw new RuntimeException("Could not find suitable preview size.");
-//        }
-//        com.google.android.gms.common.images.Size pictureSize = sizePair.pictureSize();
-//        com.google.android.gms.common.images.Size size = sizePair.previewSize();
-//        mCameraParameters.setPreviewSize(size.getWidth(), size.getHeight());
-//        mCameraParameters.setPreviewFormat(ImageFormat.NV21);
-//        if (pictureSize != null) {
-//            mCameraParameters.setPictureSize(pictureSize.getWidth(), pictureSize.getHeight());
-//        }
 
         mCamera.setParameters(mCameraParameters);
 
@@ -1167,153 +1098,6 @@ public class Camera1 extends CameraImpl {
             }
         } else {
             return normalized;
-        }
-    }
-
-    /**
-     * This runnable controls access to the underlying receiver, calling it to process frames when
-     * available from the camera.  This is designed to run detection on frames as fast as possible
-     * (i.e., without unnecessary context switching or waiting on the next frame).
-     * <p/>
-     * While detection is running on a frame, new frames may be received from the camera.  As these
-     * frames come in, the most recent frame is held onto as pending.  As soon as detection and its
-     * associated processing are done for the previous frame, detection on the mostly recently
-     * received frame will immediately start on the same thread.
-     */
-    private class FrameProcessingRunnable implements Runnable {
-        private Detector<?> mDetector;
-        private long mStartTimeMillis = android.os.SystemClock.elapsedRealtime();
-
-        // This lock guards all of the member variables below.
-        private final Object mLock = new Object();
-        private boolean mActive = true;
-
-        // These pending variables hold the state associated with the new frame awaiting processing.
-        private long mPendingTimeMillis;
-        private int mPendingFrameId = 0;
-        private java.nio.ByteBuffer mPendingFrameData;
-
-        FrameProcessingRunnable(Detector<?> detector) {
-            mDetector = detector;
-        }
-
-        /**
-         * Releases the underlying receiver.  This is only safe to do after the associated thread
-         * has completed, which is managed in camera source's release method above.
-         */
-        @android.annotation.SuppressLint("Assert")
-        void release() {
-            assert (mProcessingThread.getState() == java.lang.Thread.State.TERMINATED);
-            mDetector.release();
-        }
-
-        /**
-         * Marks the runnable as active/not active.  Signals any blocked threads to continue.
-         */
-        void setActive(boolean active) {
-            synchronized (mLock) {
-                mActive = active;
-                mLock.notifyAll();
-            }
-        }
-
-        /**
-         * Sets the frame data received from the camera.  This adds the previous unused frame buffer
-         * (if present) back to the camera, and keeps a pending reference to the frame data for
-         * future use.
-         */
-        void setNextFrame(byte[] data, Camera camera) {
-            synchronized (mLock) {
-                if (mPendingFrameData != null) {
-                    camera.addCallbackBuffer(mPendingFrameData.array());
-                    mPendingFrameData = null;
-                }
-
-                if (!mBytesToByteBuffer.containsKey(data)) {
-                    Log.d(TAG,
-                            "Skipping frame.  Could not find ByteBuffer associated with the image " +
-                                    "data from the camera.");
-                    return;
-                }
-
-                // Timestamp and frame ID are maintained here, which will give downstream code some
-                // idea of the timing of frames received and when frames were dropped along the way.
-                mPendingTimeMillis = android.os.SystemClock.elapsedRealtime() - mStartTimeMillis;
-                mPendingFrameId++;
-                mPendingFrameData = mBytesToByteBuffer.get(data);
-
-                // Notify the processor thread if it is waiting on the next frame (see below).
-                mLock.notifyAll();
-            }
-        }
-
-        /**
-         * As long as the processing thread is active, this executes detection on frames
-         * continuously.  The next pending frame is either immediately available or hasn't been
-         * received yet.  Once it is available, we transfer the frame info to local variables and
-         * run detection on that frame.  It immediately loops back for the next frame without
-         * pausing.
-         * <p/>
-         * If detection takes longer than the time in between new frames from the camera, this will
-         * mean that this loop will run without ever waiting on a frame, avoiding any context
-         * switching or frame acquisition time latency.
-         * <p/>
-         * If you find that this is using more CPU than you'd like, you should probably decrease the
-         * FPS setting above to allow for some idle time in between frames.
-         */
-        @Override
-        public void run() {
-            Frame outputFrame;
-            java.nio.ByteBuffer data;
-
-            while (true) {
-                synchronized (mLock) {
-                    while (mActive && (mPendingFrameData == null)) {
-                        try {
-                            // Wait for the next frame to be received from the camera, since we
-                            // don't have it yet.
-                            mLock.wait();
-                        } catch (InterruptedException e) {
-                            return;
-                        }
-                    }
-
-                    if (!mActive) {
-                        // Exit the loop once this camera source is stopped or released.  We check
-                        // this here, immediately after the wait() above, to handle the case where
-                        // setActive(false) had been called, triggering the termination of this
-                        // loop.
-                        return;
-                    }
-
-                    outputFrame = new Frame.Builder()
-                            .setImageData(mPendingFrameData, mPreviewSize.getWidth(),
-                                    mPreviewSize.getHeight(), android.graphics.ImageFormat.NV21)
-                            .setId(mPendingFrameId)
-                            .setTimestampMillis(mPendingTimeMillis)
-                            .setRotation(0)
-                            .build();
-
-                    // Hold onto the frame data locally, so that we can use this for detection
-                    // below.  We need to clear mPendingFrameData to ensure that this buffer isn't
-                    // recycled back to the camera before we are done using that data.
-                    data = mPendingFrameData;
-                    mPendingFrameData = null;
-                }
-
-                // The code below needs to run outside of synchronization, because this will allow
-                // The code below needs to run outside of synchronization, because this will allow
-                // the camera to add pending frame(s) while we are running detection on the current
-                // frame.
-
-                try {
-                    mDetector.receiveFrame(outputFrame);
-                } catch (Throwable t) {
-                    Log.e(TAG, "Exception thrown from receiver.", t);
-                } finally {
-                    mCamera.addCallbackBuffer(data.array());
-                }
-            }
         }
     }
 }
