+++ camerakit/src/main/base/com/wonderkiln/camerakit/CameraImpl.java
@@ -2,6 +2,9 @@ package com.wonderkiln.camerakit;
 
 import android.support.annotation.Nullable;
 
+import com.google.android.gms.vision.Detector;
+import com.google.android.gms.vision.text.TextBlock;
+
 import java.io.File;
 
 abstract class CameraImpl {
@@ -24,6 +27,7 @@ abstract class CameraImpl {
     abstract void setFocus(@Focus int focus);
     abstract void setMethod(@CaptureMethod int method);
     abstract void setZoom(@Zoom int zoom);
+    abstract void setTextDetector(Detector<TextBlock> detector);
     abstract void setVideoQuality(@VideoQuality int videoQuality);
     abstract void setVideoBitRate(int videoBitRate);
